Add files via upload
FpsNJunkCleaner is a collection of PowerShell scripts designed to improve the performance and reliability of your computer system. The FPS optimizer is implemented through a variety of performance-enhancing techniques such as closing unnecessary background programs, adjusting power plans, and updating graphics drivers.

The Junk Cleaner is a script that is specifically designed to fix slow potato PCs that are causing performance issues. It is designed to remove unnecessary files, folders, and programs that can slow down your computer, freeing up valuable resources and increasing overall performance.

With FpsNJunkCleaner, you can expect a noticeable improvement in your computer's performance by running the provided scripts. It is the perfect tool for anyone looking to get the most out of their potato PC and enhance their gaming experience. Please keep in mind that it's necessary to have PowerShell installed on your system in order to run the scripts and to be familiar with the basic usage of it.